+++ build.mac.sh
@@ -30,6 +30,22 @@ xcodebuild -project HiveMP.ClientConnect.xcodeproj -configuration Release build
 
 cd $ROOT
 
+echo "Testing 32-bit binaries..."
+cd $ROOT/buildmac32_${SHASUM}_v1/Release
+./HiveMP.SteamTest | tee result.txt
+if [ "$(cat result.txt | tr " " "\n" | grep -c "TEST PASS")" != "2" ]; then
+    echo "Test failed!"
+    exit 1
+fi
+
+echo "Testing 64-bit binaries..."
+cd $ROOT/buildmac64_${SHASUM}_v1/Release
+./HiveMP.SteamTest | tee result.txt
+if [ "$(cat result.txt | tr " " "\n" | grep -c "TEST PASS")" != "2" ]; then
+    echo "Test failed!"
+    exit 1
+fi
+
 echo "Creating distribution structure..."
 if [ -d dist ]; then
     rm -Rf dist
